Added HTML, Javascript, and CSS
Added the code to make it so when you press the calculate button the answer will be printed to the html. Centered everything on the page, added margins so the paragraph looks better, and I moved the title down a bit. I also moved the calculate button and text box to below the paragraph that explains what the website does.